refactor: move more common code into ModuleHelpers
