+++ MiraBot.Miraminders/MiraminderService.cs
@@ -33,12 +33,6 @@ namespace MiraBot.Miraminders
             return user?.Timezone;
         }
 
-        public static bool IsValidTimezone(string timezoneId)
-        {
-            return TimeZoneInfo
-                .GetSystemTimeZones()
-                .Any(t => t.Id.Equals(timezoneId, StringComparison.OrdinalIgnoreCase));
-        }
 
         public async Task AddReminderAsync(Reminder reminder)
         {
@@ -85,35 +79,6 @@ namespace MiraBot.Miraminders
             return userTime;
         }
 
-
-        public async Task AddTimezoneToUserAsync(ulong discordId, string timezoneId)
-        {
-            var user = await _usersRepository.GetUserByDiscordIdAsync(discordId)
-                .ConfigureAwait(false);
-
-            if (user is not null)
-            {
-                user.Timezone = timezoneId;
-                await _usersRepository.ModifyUserAsync(user);
-            }
-        }
-
-        public async Task AddDateFormatToUserAsync(ulong discordId, bool isAmerican)
-        {
-            var user = await _usersRepository.GetUserByDiscordIdAsync(discordId)
-                .ConfigureAwait(false);
-
-            user.UsesAmericanDateFormat = isAmerican;
-            await _usersRepository.ModifyUserAsync(user);
-        }
-
-        public static string CreateTimezoneFile()
-        {
-            var fileName = Path.ChangeExtension(Path.GetRandomFileName(), ".txt");
-            var timeZones = TimeZoneInfo.GetSystemTimeZones();
-            File.WriteAllLines(fileName, timeZones.Select(t => t.Id).ToArray());
-            return fileName;
-        }
 
         public async Task<List<string>> SendLongMessage(List<Reminder> reminders)
         {
